OVS-agent: Separate ovs-ofctl using code as a driver
This is a preparation to introduce another Ryu-based driver.
The aim is to replace this with the new Ryu-based driver eventually.

Also, this commit simplifies DVR logics by reducing duplications
and makes some of DVR UTs actually check the flows rather than just
"add_flow is called".

Partially-Implements: blueprint ovs-ofctl-to-python
Change-Id: Ie1224f8a1c17268cd7d1c474ed82fdfb8852eaa8